==> 2_Twitter_Data_Mining.ipynb <==
# Reading the Json Files & Data Mining

The tweets were saved on 73 json files with sizes ranging from 100 MB to ~19 GBs. All the files were broken to pieces of 100,000 lines (~ 700 MB) so that they can be easily processed on my PC.
I process the json files individually and save the resulting dataframes in CSV format. These CSV files are significantly smaller (~ 3-8 MB) and will be concatenated later.

I have cetegorized the tweets to tweets and retweets and saved them in separate CSV files. Retweet dataframes contain information about the retweet and the original tweet. The final tweet csv file is ~570 MB and contains ~1.7 million tweets. The final retweet csv file is ~1.4 GB and contains ~2.9 million retweets. 


import simplejson as json
import pandas as pd
from glob import glob
import numpy as np


# function to extract user_mentions from user_mentions list of dictionaries
userextract = lambda x: [y['screen_name'] for y in x] if len(x)> 0 else []
# function to extract hashtags from hashtags list of dictionaries
hashextract = lambda x: [y['text'] for y in x] if len(x)> 0 else []

# function to extract extended state, user-mentions, hashtags, and full-texts
def multi_functions(tweet):
    if 'extended_tweet' in tweet.keys():
        return [ True, userextract(tweet['extended_tweet']['entities']['user_mentions']), \
                hashextract(tweet['extended_tweet']['entities']['hashtags']), tweet['extended_tweet']['full_text']]
    else:
        return [False, userextract(tweet['entities']['user_mentions']),\
                hashextract(tweet['entities']['hashtags']), tweet['text']]

# csv file to keep track of the number of the duplicate tweets, change 1000 if you have more than 1000 files  
tweet_counter = pd.DataFrame(index=np.arange(1000), \
                             columns=['filename', 'recorded_tweets', 'recorded_retweets', 'tweet_counts', 'retweet_count'])

# using glob to read the file names
filenames = glob('data/blackfriday*.json')
# or your own list
#filenames = ['blackfriday46_0.json']

i = 0
tweets_data = []
retweets_data = []
data = []

# reading the json files
for filename in filenames:
    tweets_file = open(filename, "r")
    for line in tweets_file:
        try:
            tweet = json.loads(line)
            data.append(tweet)
        except:
            continue

    tweets_file.close()

    #############################################################################################

    # dividing the tweets to tweets & retweets categories
    for tweet in data:
        if 'text' in tweet.keys(): # filter error lines 
            if 'retweeted_status' in tweet.keys():
                retweets_data.append(tweet)
            else:
                tweets_data.append(tweet)
          
    data = []
    
    tweet_counter.iloc[i, 0] = filename
    tweet_counter.iloc[i, 1] = len(tweets_data)
    tweet_counter.iloc[i, 2] = len(retweets_data)

    #############################################################################################

    # building a tweets dataframe & saving it in csv format
    tweets = pd.DataFrame()
    tweets['id'] = [tweet['id'] for tweet in tweets_data]
    tweets['created_at'] = [tweet['created_at'] for tweet in tweets_data]
    tweets['language'] = [tweet['lang'] for tweet in tweets_data]
    tweets['coordinates'] = [tweet['coordinates'] for tweet in tweets_data]

    tweets['mixed'] = [multi_functions(tweet) for tweet in tweets_data]
    tweets['extended'] = tweets['mixed'].apply(lambda x: x[0])
    tweets['user_mentions'] = tweets['mixed'].apply(lambda x: x[1])
    tweets['hashtags'] = tweets['mixed'].apply(lambda x: x[2])
    tweets['full_text'] = tweets['mixed'].apply(lambda x: x[3])
    tweets['full_text'] = tweets['full_text'].apply(lambda x: [x])
    tweets.drop('mixed', axis=1, inplace=True)

    tweets['user_name'] = [tweet['user']['screen_name'] for tweet in tweets_data]
    tweets['verified'] = [ True if tweet['user']['verified'] is True else False for tweet in tweets_data]
    tweets['user_followers'] = [tweet['user']['followers_count'] for tweet in tweets_data]
    tweets['user_friends'] = [tweet['user']['friends_count'] for tweet in tweets_data]
    tweets['user_favourites_count'] = [tweet['user']['favourites_count'] for tweet in tweets_data]
    tweets['user_listed_count'] = [tweet['user']['listed_count'] for tweet in tweets_data]
    tweets['user_statuses_count'] = [tweet['user']['statuses_count'] for tweet in tweets_data]
    tweets['user_created_at'] = [tweet['user']['created_at'] for tweet in tweets_data]
    tweets['retweet'] = False
    
    tweets = tweets.drop_duplicates('id')
    
    filename_tweets = 'output/tweets'+str(i)
    tweets.to_csv(filename_tweets, index=False)
    tweet_counter.iloc[i, 3] = len(tweets)
    tweets_data = []
    tweets = pd.DataFrame()
    
    #############################################################################################
    
    # building a retweets dataframe & saving it in csv format
    retweets = pd.DataFrame()
    retweets['id'] = [tweet['id'] for tweet in retweets_data]
    retweets['created_at'] = [tweet['created_at'] for tweet in retweets_data]
    retweets['language'] = [tweet['lang'] for tweet in retweets_data]
    retweets['coordinates'] = [tweet['coordinates'] for tweet in retweets_data]

    retweets['mixed'] = [multi_functions(tweet['retweeted_status']) for tweet in retweets_data]
    retweets['extended'] = retweets['mixed'].apply(lambda x: x[0])
    retweets['user_mentions'] = retweets['mixed'].apply(lambda x: x[1])
    retweets['hashtags'] = retweets['mixed'].apply(lambda x: x[2])
    retweets['full_text'] = retweets['mixed'].apply(lambda x: x[3])
    retweets['full_text'] = retweets['full_text'].apply(lambda x: [x])
    retweets.drop('mixed', axis=1, inplace=True)

    retweets['user_name'] = [tweet['user']['screen_name'] for tweet in retweets_data]
    retweets['user_followers'] = [tweet['user']['followers_count'] for tweet in retweets_data]
    retweets['user_friends'] = [tweet['user']['friends_count'] for tweet in retweets_data]
    retweets['user_favourites_count'] = [tweet['user']['favourites_count'] for tweet in retweets_data]
    retweets['user_listed_count'] = [tweet['user']['listed_count'] for tweet in retweets_data]
    retweets['user_statuses_count'] = [tweet['user']['statuses_count'] for tweet in retweets_data]
    retweets['user_created_at'] = [tweet['user']['created_at'] for tweet in retweets_data]

    retweets['retweet'] = True
    retweets['orig_id'] = [tweet['retweeted_status']['id'] for tweet in retweets_data]
    retweets['orig_created_at'] = [tweet['retweeted_status']['created_at'] for tweet in retweets_data]
    retweets['orig_user'] = [tweet['retweeted_status']['user']['screen_name'] for tweet in retweets_data]
    retweets['verified'] = [ True if tweet['retweeted_status']['user']['verified'] is True else False for tweet in retweets_data]
    retweets['orig_user_followers'] = [tweet['retweeted_status']['user']['followers_count'] for tweet in retweets_data]
    retweets['orig_user_friends'] = [tweet['retweeted_status']['user']['friends_count'] for tweet in retweets_data]
    retweets['orig_user_favourites'] = [tweet['retweeted_status']['user']['favourites_count'] for tweet in retweets_data]
    retweets['orig_user_listed'] = [tweet['retweeted_status']['user']['listed_count'] for tweet in retweets_data]
    retweets['orig_user_statuses'] = [tweet['retweeted_status']['user']['statuses_count'] for tweet in retweets_data]
    retweets['orig_user_created_at'] = [tweet['retweeted_status']['user']['created_at'] for tweet in retweets_data]

    retweets = retweets.drop_duplicates('id') 

    filename_retweets = 'output/retweets'+str(i)
    retweets.to_csv(filename_retweets, index=False)
    tweet_counter.iloc[i, 4] = len(retweets)
    retweets_data = []
    retweets = pd.DataFrame()
    
    # print(i)
    i +=1

# saving the tweet counter data
tweet_counter.to_csv('tweet_counter.csv')


# Concatanation
*This section can be run independent of the above section*

import pandas as pd
from glob import glob

filenames_tweets = glob('output/tweets*')

list_tweets = []

for filename in filenames_tweets:
    df_chunk = pd.read_csv(filename)
    list_tweets.append(df_chunk)

# Free up the memory
df_chunk = pd.DataFrame()

df_tweets = pd.concat(list_tweets, axis= 0)
df_tweets = df_tweets.drop_duplicates('id')

list_tweets = []
df_tweets.to_csv('df_tweets.csv', index=False)

df_tweet.head()

# Free up the memory if you need to
df_tweets = pd.DataFrame()

filenames_retweets = glob('output/retweets*')

list_retweets = []

for filename in filenames_retweets:
    df_chunk = pd.read_csv(filename)
    list_retweets.append(df_chunk)

# Free up the memory
df_chunk = pd.DataFrame()

df_retweets = pd.concat(list_retweets, axis= 0)
df_retweets = df_retweets.drop_duplicates('id')

list_retweets = []
df_retweets.to_csv('df_retweets.csv', index=False)

df_retweets.head()

# Free up the memory if you need to
df_retweets = pd.DataFrame()
