==> 5_Twitter_Sentiment_Analysis.ipynb <==
# Tweet Sentiment Analysis

*Simple sentiment analysis using text blob*

Results for the whole Tweet dataset (English): 51% positive sentiment, 36% neutral sentiment, 13% negative sentiment.

- Results for the tweets with #apple or @apple: 52% positive sentiment, 41% neutral sentiment, 7% negative sentiment.

- Results for the tweets with #samsung or @samsung: 50% positive sentiment, 37% neutral sentiment, 13% negative sentiment.

It seems that text blob False Negative rate is not trivial for the tweets about shopping, deals and discounts. However, we can still compare the sentiment about different companies with each other, assuming that the False Negative rate is the same for large samples. Relatively speaking, the sentiment of tweets about Apple around Black Friday & Cyber Monday was more positive than the sentiment of tweets about Samsung.

Sentiment Analysis can help businesses in early detection of negative tweets/comments and addressing customer concerns.

import pandas as pd
import numpy as np

tweets = pd.read_csv('df_tweets.csv', usecols=['created_at', 'language', 'full_text', 'hashtags', 'user_mentions'],\
                     parse_dates=['created_at'], infer_datetime_format=True, low_memory=False)

# drop the tweets in languages other than English 
tweets = tweets[tweets['language'] == 'en']

print (len(tweets), 'tweets out of the ~1.7 million collected tweets were in english')

tweets.head()

# Sentiment Analysis 
from textblob import TextBlob
import re

def clean_tweet(tweet):
    '''
    Utility function to clean the text in a tweet by removing 
    links and special characters using regex.
    '''
    return ' '.join(re.sub("(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)", " ", tweet).split())

def analize_sentiment(tweet):
    '''
    Utility function to classify the polarity of a tweet
    using textblob.
    '''
    analysis = TextBlob(clean_tweet(tweet))
    if analysis.sentiment.polarity > 0:
        return 1
    elif analysis.sentiment.polarity == 0:
        return 0
    else:
        return -1

# Adding a column for sentiment    
tweets['SA'] = tweets['full_text'].apply(lambda x: analize_sentiment(x))

# Percentages
sentiments = tweets.groupby('SA').agg({'SA':'count'}).rename(columns=({'SA': 'Count'}))
sentiments['% of total'] = sentiments['Count']/sentiments['Count'].sum()
sentiments ['Sentiment'] = ['Negative', 'Neutral', 'Positive']
sentiments = sentiments.reindex(columns = ['Sentiment', 'Count','% of total'])
sentiments

## Tweet Sentiments: Apple vs Samsung

# Changing the type of the hashtags and user_mentions columns from string to list
tweets['hashtags'] = tweets['hashtags'].apply(lambda x: x.strip('[]').lower().replace("'",'').split(', '))
tweets['user_mentions'] = tweets['user_mentions'].apply(lambda x: x.strip('[]').lower().replace("'",'').split(', '))

# tweets with hashtags or user mentions of apple or samsung
tweets ['apple'] = (tweets['hashtags'] + tweets['user_mentions']).apply(lambda x: True if 'apple' in x else False)
tweets ['samsung'] = (tweets['hashtags'] + tweets['user_mentions']).apply(lambda x: True if 'samsung' in x else False)

apple = tweets[tweets['apple'] == True]
apple.head()

samsung = tweets[tweets['samsung'] == True]
samsung.head()

# Sentiment around Apple
apple_SA = apple.groupby('SA').agg({'SA':'count'}).rename(columns=({'SA': 'Count'}))
apple_SA ['% of total'] = apple_SA['Count']/apple_SA['Count'].sum()
apple_SA ['Sentiment'] = ['Negative', 'Neutral', 'Positive']
apple_SA = apple_SA.reindex(columns = ['Sentiment', 'Count','% of total'])
apple_SA

# Sentiment around Samsung
samsung_SA = samsung.groupby('SA').agg({'SA':'count'}).rename(columns=({'SA': 'Count'}))
samsung_SA ['% of total'] = samsung_SA['Count']/samsung_SA['Count'].sum()
samsung_SA ['Sentiment'] = ['Negative', 'Neutral', 'Positive']
samsung_SA = samsung_SA.reindex(columns = ['Sentiment', 'Count','% of total'])
samsung_SA

# Example of a tweet with nagative sentiment about samsung (True Negative)
samsung['full_text'][4309]

# Example of a tweet with detected nagative sentiment about samsung which is not actually negative (False Negative)
samsung['full_text'][1210]
